requirements.txt
flask==2.0.1
Werkzeug==2.0.1
numpy==1.21.0
pillow==8.3.1
scikit-learn==1.0.2
gunicorn==21.0.0